+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.0] - 2024-02-22
+
+### Added
+- Add support for using unquantized models in the GGUF format from the source
+
 ## [1.2.0] - 2024-02-20
 
 ### Added